Replaced the base code and added base cube
Changed the base code to my last project. I stripped out almost
all of the old functions and objects, replacing them with a basecube.
changed the way objects are drawn so they will not resubmit the
vertices every time. Also set it to have a transform matrix to move
and rotate the object.