Added the first edible plant
I named the plant a pumpkin, but the color of the fruit turned out
to be a pinkish color which I liked so I kept it. The plant goes throug
a short growing animation that stops when it is fully grown, it reverts
back to its starting stage when it has been harvested.